+++ Runtime/Audio/AudioSourcePool.cs
@@ -1,8 +1,18 @@
-using UnityEngine;
+﻿using UnityEngine;
 namespace Etienne.Pools
 {
     public class AudioSourcePool : ComponentPool<AudioSource>
     {
+	    public static AudioSource Play(Cue cue, int index, Vector3 position)
+        {
+            if (index < 0 || index >= cue.Clips.Length)
+            {
+                Debug.LogWarning("Index out of range, played random sound instead.");
+                return cue.Sound.Play(position);
+            }
+            return new Sound(cue.Clips[index], cue.Parameters).Play(position);
+        }
+        
         public static AudioSource Play(Cue cue, int index, Transform transform = null)
         {
             if (index < 0 || index >= cue.Clips.Length)
@@ -11,6 +21,18 @@ namespace Etienne.Pools
                 return cue.Sound.Play(transform);
             }
             return new Sound(cue.Clips[index], cue.Parameters).Play(transform);
+        }
+
+	    public static AudioSource Play(Sound sound, Vector3 position)
+        {
+            if (instance == null) CreateInstance<AudioSourcePool>(100);
+
+            AudioSource source = instance.Dequeue();
+            source.SetSoundToSource(sound);
+            source.Play();
+	        source.transform.position= position;
+            EnqueueSoundAfterClip(source);
+            return source;
         }
 
         public static AudioSource Play(Sound sound, Transform transform = null)
@@ -35,6 +57,18 @@ namespace Etienne.Pools
             if (!Application.isPlaying || source == null) return;
 
             instance.Enqueue(source);
+        }
+
+	    public static AudioSource PlayLooped(Sound sound, Vector3 position)
+        {
+            if (instance == null) CreateInstance<AudioSourcePool>(100);
+
+            AudioSource source = instance.Dequeue();
+            source.SetSoundToSource(sound);
+            source.loop = true;
+            source.Play();
+            source.transform.position = position;
+            return source;
         }
 
         public static AudioSource PlayLooped(Sound sound, Transform transform = null)